Menu & MenuOption 생성 기능 수정

+++ src/main/java/net/softsociety/bacs/menu/dto/UpdateMenuDTO.java
@@ -12,7 +12,7 @@ public record UpdateMenuDTO(
         @NotBlank
         String menuName,
         @NotBlank
-        long menuPrice,
+        int menuPrice,
         @NotBlank
         String menuImg,
         @NotBlank
